feat: 스마트스토어 상품주문 연동 팝업 & JSON 전송
- 브라우저 창 가운데에 팝업을 표시하는 함수 구현
- 상품주문 연동을 위한 JSON 요청 구현
