+++ src/main/java/kr/tatine/manibogo_oms_v2/common/model/Money.java
@@ -19,4 +19,8 @@ public class Money {
     public Money(BigDecimal value) {
         this.value = value;
     }
+
+    public Money(long value) {
+        this(BigDecimal.valueOf(value));
+    }
 }
