+++ src/main/resources/templates/fragments/layout.html
@@ -19,10 +19,11 @@
       <div th:replace="${content}">레이아웃 내용</div>
     </body>
 
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/babel-polyfill/6.26.0/polyfill.min.js" integrity="sha512-vDTLyifjBBdmGeJrJMO7w+qbbk+7uyoKsUxkhgxPtn3YShSVspezU0EXf780txBgNXzfKkEt7a3RhLs7NmInJQ==" crossorigin="anonymous" referrerpolicy="no-referrer"></script>
     <script lang="javascript" src="https://cdn.sheetjs.com/xlsx-0.20.3/package/dist/xlsx.full.min.js"></script>
 
     <script src="/js/table.js"></script>
+    <script src="/js/popup.js"></script>
+
     <th:block th:replace="${scripts}">레이아웃 추가 스크립트</th:block>
 
 </html>
